+++ packages/browser-extension/src/hooks/tlsnProve/redaction/index.ts
@@ -1,0 +1,1 @@
+export * from "./redact";
