+++ packages/browser-extension/src/hooks/tlsnProve/redaction/urlQuery/request.query.ranges.ts
@@ -1,4 +1,4 @@
-import { Utf8String } from "./utils";
+import { Utf8String } from "../utils";
 
 const calculateRequestQueryParamsRanges = (
   urlQueryParams: string[],
