+++ packages/browser-extension/src/hooks/tlsnProve/redaction/utils/error.ts
